+++ Patterns/Composite/App/compose/Models/Concrete/RobotChassis.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace compose.Models.Concrete
+{
+  public class RobotChassis
+  {
+    public int? ArmCount { get; set; }
+
+    public string GetGuid()
+    {
+      return Guid.NewGuid().ToString();
+    }
+  }
+}
